commit 21 marzo PF

+++ ws_programacion2/ws_programacion2Spring/drogueria-api/drogueria-api/drogueria-api/src/main/java/com/drogueria/api/services/ClienteServiceImpl.java
@@ -8,14 +8,18 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.drogueria.api.entities.Cliente;
+import com.drogueria.api.entities.Domicilio;
 import com.drogueria.api.exception.ClienteException;
 import com.drogueria.api.repositories.IClienteRepository;
+import com.drogueria.api.repositories.IDomicilioRepository;
 
 @Service
 public class ClienteServiceImpl implements IClienteService{
 
 	@Autowired
 	IClienteRepository clienteRepository;
+	@Autowired
+	IDomicilioRepository domicilioRepository;
 	
 	@Override
 	@Transactional
@@ -26,7 +30,13 @@ public class ClienteServiceImpl implements IClienteService{
 
 	@Override
 	@Transactional
-	public void save(Cliente cliente) {
+	public void save(Cliente cliente) throws ClienteException{
+		Cliente cliente2= null;
+		cliente2=clienteRepository.findClienteByCedula(cliente.getCedula());
+		
+		if (cliente2!=null) {
+			throw new ClienteException("El cliente ya esta registrado con esta cedula");
+		}
 		clienteRepository.save(cliente);
 	}
 
@@ -47,7 +57,13 @@ public class ClienteServiceImpl implements IClienteService{
 
 	@Override
 	@Transactional
-	public void delete(int id) {
+	public void delete(int id) throws ClienteException{
+		Domicilio domicilio= null;
+		domicilio=domicilioRepository.findClienteByIdEliminar(id);
+		
+		if (domicilio!=null) {
+			throw new ClienteException("El cliente no se puede eliminar");
+		}
 		clienteRepository.deleteById(id);
 	}
